requirements.txt
fastapi
uvicorn
pygerber
Pillow